refactor(timer):重构超时定时器管理逻辑
- 将 CountDownTimer 相关代码提取到独立的 TimeoutTimerManager 类中
- 移除 MainActivity 中的 timeoutTimer 字段和相关方法
- 在 FloatingWindowService 中将 tick 参数从 Long 类型改为 Int 类型
- 统一超时定时器的生命周期管理
- 优化定时器启动和销毁逻辑，避免内存泄漏风险
- 简化倒计时更新的广播发送机制

+++ app/src/main/java/com/pengxh/daily/app/service/FloatingWindowService.kt
@@ -61,7 +61,7 @@ class FloatingWindowService : Service() {
 
                     MessageType.UPDATE_FLOATING_WINDOW_TIME -> {
                         // 更新悬浮窗倒计时
-                        val tick = intent.getLongExtra("tick", 30)
+                        val tick = intent.getIntExtra("tick", 30)
                         binding.timeView.text = "${tick}s"
                         if (tick < 1) {
                             binding.root.alpha = 0.0f
